test: wildcard test cases with and without disclosure provision by the wallet, refactor presentation definition policy code to main test package
